+++ R/getimg.R
@@ -2,7 +2,7 @@
 #'
 #' Finds and imports PNG, JPEG, and/or BMP images.
 #'
-#' @param imgpath (required) either the full path to a given image (including extension),
+#' @param imgpath (required) eith the full file-path or URL to an image (including extension),
 #' or the path to a folder in which multiple image files are located. Mixed file formats
 #' within a folder are accepted.
 #' @param subdir should subdirectories within the \code{imgpath} folder be
@@ -11,16 +11,14 @@
 #' images? (defaults to \code{FALSE}).
 #' @param max.size maximum size of all images to be allowed in memory, in GB. Defaults to
 #' \code{1}.
-#' @param cores number of cores to be used in parallel processing. If \code{1}, or
-#' if total image sizes exceed 200 mb in memory, parallel computing will not be used.
-#' Defaults to \code{getOption("mc.cores", 2L)}. Not available on Windows.
+#' @param cores deprecated argument.
 #'
 #' @return a image, or list of images, of class \code{rimg},
 #' for use in further \code{pavo} functions.
 #'
-#' @export
+#' @importFrom magick image_info
 #'
-#' @importFrom pbmcapply pbmclapply
+#' @export
 #'
 #' @examples
 #' # Single image
@@ -28,10 +26,11 @@
 #' 
 #' # Multiple images
 #' snakes <- getimg(system.file("testdata/images/snakes", package = "pavo"))
+#' 
 #' @author Thomas E. White \email{thomas.white026@@gmail.com}
 
 getimg <- function(imgpath = getwd(), subdir = FALSE, subdir.names = FALSE,
-                   max.size = 1, cores = getOption("mc.cores", 2L)) {
+                   max.size = 1, cores) {
 
   ## ------------------------------ Checks ------------------------------ ##
 
@@ -39,21 +38,20 @@ getimg <- function(imgpath = getwd(), subdir = FALSE, subdir.names = FALSE,
   ext <- c("jpg", "jpeg", "png", "bmp")
 
   ## Cores
-  if (cores > 1 && .Platform$OS.type == "windows") {
-    cores <- 1
+  if (!missing(cores)) {
+    warning("the cores argument is deprecated as all image importing is now vectorised.",
+      call. = FALSE
+    )
   }
 
   ## ------------------------------ Main ------------------------------ ##
 
   # If file extensions are in 'imgpath', it's a single image being directly specified
   if (grepl(paste(ext, collapse = "|"), imgpath, ignore.case = TRUE)) {
-    imgdat <- grabimg(imgpath)
-    imgdat <- as.rimg(drop(as.array(imgdat)),
-      name = sub(".*\\/", "", sub("[.][^.]+$", "", imgpath))
-    )
+    imgdat <- as.rimg(image_read(imgpath), name = sub(".*\\/", "", sub("[.][^.]+$", "", imgpath)))
 
     # Warn of slowness if dimensions are large
-    if ((dim(imgdat)[1] * dim(imgdat)[2]) > (1000 * 1000)) {
+    if (dim(imgdat)[1] * dim(imgdat)[2] > 1000000) {
       message("Image dimensions are relatively large, consider reducing image size with procimg() for faster performance.")
     }
 
@@ -85,7 +83,7 @@ getimg <- function(imgpath = getwd(), subdir = FALSE, subdir.names = FALSE,
     imgnames <- gsub(extension, "", file_names)
 
     # Stop if max size estimated to exceed available memory
-    imgsize <- prod(dim(grabimg(files[1])))
+    imgsize <- image_info(image_read(files[1]))["filesize"]
     totalsize <- ((imgsize * 8) * length(file_names)) / (1024^3)
     if (totalsize > max.size) {
       stop("Total size of images likely exceeds available memory. Check max.size is set appropriately.")
@@ -96,39 +94,13 @@ getimg <- function(imgpath = getwd(), subdir = FALSE, subdir.names = FALSE,
       message("Total size of images exceeds 200 mb in memory, which may result in slowed performance. Consider resizing images with procimg() prior to analysis, if speed is a priority.")
     }
 
-    # Crudely avoid a bug in pbmclapply when handling large objects.
-    if (totalsize < 0.1) {
-      imgdat <- pbmclapply(files, grabimg, mc.cores = cores)
-    } else {
-      imgdat <- lapply(files, grabimg)
-    }
+    # Get images
+    imgdat <- as.rimg(image_read(files), name = imgnames)
 
-    imgdat <- lapply(imgdat, function(x) drop(as.array(x)))
-    imgdat <- as.rimg(imgdat, imgnames)
-
-    # Simplify if it's a single image
-    if (length(imgdat) == 1) imgdat <- cimg2rimg(imgdat[[1]])
+    # Simplify if it's a single image   ###TODO###
+    if (length(imgdat) == 1) imgdat <- imgdat[[1]]
   }
   imgdat
-}
-
-## Grab images
-#' @importFrom readbitmap read.bitmap
-grabimg <- function(file) {
-  bmp <- read.bitmap(file)
-  if (!is.null(attr(bmp, "header"))) {
-    bmp <- bmp / 255
-  }
-  if (length(dim(bmp)) == 3) { # 3 channels (colour)
-    bmp <- mirrorx(bmp)
-    bmp <- rot90(bmp)
-  }
-  else { # 1 channel (B&W), duplicate to 3d for classification convenience
-    bmp <- replicate(3, bmp, simplify = "array")
-    bmp <- mirrorx(bmp)
-    bmp <- rot90(bmp)
-  }
-  bmp
 }
 
 ## Rotate matrices 90-degrees